fix(auth): add /auth/callback route, set OAuth redirectTo, SPA fallback
